services/users/requirements.txt
Flask==1.0.2
python-dotenv==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.7
Flask-Testing==0.7.1
gunicorn==19.9.0
coverage==4.5.3
flake8===3.7.7
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
flask-migrate==2.4.0
